New: "Update ItemController.Delete Action"
Update the ItemController.Delete action to prevent deleting items by anyone but the user who owns that item.

Before removing the Item check to make sure the Item's User is the same as the logged in User, if not return Unauthorized

+++ WishList/Controllers/ItemController.cs
@@ -46,6 +46,7 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if (item.User.Id != _userManager.GetUserAsync(HttpContext.User).Result.Id) return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
